# packages/shared-types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Authenticate user (email + password) — issue session cookies",
        "description": "Returns 200 with 2 Set-Cookie headers (icp_session SameSite=Lax + icp_refresh SameSite=Strict) and body {user: PublicUser}. NO tokens in body per ADR-019 + S-03 C-01.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Invalidate current session",
        "description": "Requires icp_session cookie. Clears both auth cookies (Max-Age=0) + DEL Redis session:{jti} + UPDATE PG sessions.revoked_at=NOW(). Idempotent.",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "icp_session": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "summary": "Get current user profile",
        "description": "Returns 200 with user shape including avatar_initials (server-computed) and last_login_at (MAX issued_at from sessions). Requires icp_session cookie.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "icp_session": []
          }
        ]
      }
    },
    "/api/v1/auth/switch-tenant": {
      "post": {
        "operationId": "AuthController_switchTenant",
        "summary": "Đổi shop — cập nhật landing hint, trả redirect_url (KHÔNG re-issue token)",
        "description": "ADR-046 amend (c): active tenant = URL, KHÔNG phải claim. Verify membership → UPDATE sessions.last_active_tenant_id → trả { tenant_id, slug, redirect_url } để FE router.push(redirect_url). JWT/cookie KHÔNG đổi. 403 nếu không phải member.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchTenantRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwitchTenantResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of target tenant"
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "icp_session": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Rotating refresh — exchange refresh cookie for new pair",
        "description": "Reads icp_refresh cookie (NOT icp_session — separate path). Atomically revokes old refresh + issues new pair. Replay → 401. Per S-03 C-06.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Refresh token missing/expired/revoked"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "summary": "Request password reset — stub (no real SMTP)",
        "description": "Per S-03 C-03 stub Phase 02. Always returns {sent: true} regardless of email existence (no user enumeration per OWASP). Emits auth.password_reset_requested behavior event for analytics tracking. NO database query, NO SMTP integration. Phase 6 productionization adds real email service + reset token table.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "ICP Gateway API",
    "description": "Intelligent Commerce Platform — REST API.\n\n4 mutating endpoints require `Idempotency-Key` header (UUID v4) per docs/03_API_CONTRACTS.md §1.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "icp_session",
        "description": "httpOnly session cookie issued by POST /auth/login (per ADR-019 + S-03 C-01)."
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 255
          },
          "remember_me": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "role": {
                "type": "string",
                "enum": [
                  "merchant",
                  "customer",
                  "admin"
                ]
              },
              "display_name": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "avatar_initials": {
                "type": "string",
                "minLength": 1,
                "maxLength": 2
              }
            },
            "required": [
              "id",
              "email",
              "role",
              "display_name",
              "avatar_initials"
            ]
          }
        },
        "required": [
          "user"
        ]
      },
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "merchant",
              "customer",
              "admin"
            ]
          },
          "display_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "avatar_initials": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "session_expires_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "display_name",
          "avatar_initials",
          "last_login_at",
          "session_expires_at"
        ]
      },
      "SwitchTenantRequestDto": {
        "type": "object",
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "tenant_id"
        ]
      },
      "SwitchTenantResponseDto": {
        "type": "object",
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "redirect_url": {
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "slug",
          "redirect_url"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255
          }
        },
        "required": [
          "email"
        ]
      },
      "ForgotPasswordResponseDto": {
        "type": "object",
        "properties": {
          "sent": {
            "type": "boolean"
          }
        },
        "required": [
          "sent"
        ]
      },
...
    }
  }
}