# packages/shared-types/src/api/models/SwitchTenantResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SwitchTenantResponseDto = {
  tenant_id: string;
  slug: string;
  redirect_url: string;
};